fix: consolidate scripts a bit
