+++ docker-compose.yml
@@ -1,5 +1,5 @@
 x-common-vars:
-  - &STACKS_BLOCKCHAIN_COMMIT 190a195f690025a07371cb2a5e84b3f17cc63c92 # feat/epoch-3-5-rc
+  - &STACKS_BLOCKCHAIN_COMMIT adfcb143a939c8723318fc26252ba3eaba2265b3 # feat/epoch-3-5-rc
   - &STACKS_API_COMMIT cb5881749553a34704cab62a9631341cb30f2c17 # 8.13.4
   - &BITCOIN_ADDRESSES miEJtNKa3ASpA19v5ZhvbKTEieYjLpzCYT
   - &MINER_SEED 9e446f6b0c6a96cf2190e54bcd5a8569c3e386f091605499464389b8d4e0bfc201 # stx: STEW4ZNT093ZHK4NEQKX8QJGM2Y7WWJ2FQQS5C19, btc: miEJtNKa3ASpA19v5ZhvbKTEieYjLpzCYT, pub_key: 035379aa40c02890d253cfa577964116eb5295570ae9f7287cbae5f2585f5b2c7c, wif: cStMQXkK5yTFGP3KbNXYQ3sJf2qwQiKrZwR9QJnksp32eKzef1za
